requirements.txt
awsebcli==3.15.3
blessed==1.16.1
botocore==1.12.253
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
Django==2.1.1
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
jmespath==0.9.4
jsonschema==2.6.0
mysqlclient==1.4.5
pathspec==0.5.9
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
